Fixed foreign object bug.

+++ db/migrate/20190611095651_create_relationships.rb
@@ -3,7 +3,7 @@ class CreateRelationships < ActiveRecord::Migration[5.2]
     create_table   :relationships, :primary_key => :relationship_id do |t|
     	t.bigint       :person_id_a, null:false
     	t.bigint       :person_id_b, null:false
-    	t.integer      :relationship_type_id, null:false
+    	t.bigint      :relationship_type_id, null:false
       t.bigint       :creator, null:false
     	t.boolean      :voided, null: false, default: 0
     	t.bigint       :voided_by
